third_party/blink/web_tests/http/tests/inspector-protocol/prerender/prefetch-events-for-ahead-of-prerender-expected.txt


Running test: prefetchStatusUpdatedShouldBeEmittedForPrefetchAheadOfPrerender
Preload.prefetchStatusUpadted should be emitted for prefetch ahead of prerender
[
    [0] : {
        method : Preload.prefetchStatusUpdated
        params : {
            initiatingFrameId : <string>
            key : {
                action : Prefetch
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            prefetchStatus : PrefetchNotFinishedInTime
            prefetchUrl : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            requestId : <string>
            status : Running
        }
        sessionId : <string>
    }
    [1] : {
        method : Preload.prerenderStatusUpdated
        params : {
            key : {
                action : Prerender
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            status : Pending
        }
        sessionId : <string>
    }
    [2] : {
        method : Preload.prerenderStatusUpdated
        params : {
            key : {
                action : Prerender
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            status : Running
        }
        sessionId : <string>
    }
    [3] : {
        method : Preload.prefetchStatusUpdated
        params : {
            initiatingFrameId : <string>
            key : {
                action : Prefetch
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            prefetchStatus : PrefetchResponseUsed
            prefetchUrl : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            requestId : <string>
            status : Success
        }
        sessionId : <string>
    }
    [4] : {
        method : Preload.prerenderStatusUpdated
        params : {
            key : {
                action : Prerender
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            status : Ready
        }
        sessionId : <string>
    }
]
[
    [0] : {
        method : Network.requestWillBeSent
        params : {
            documentURL : <string>
            frameId : <string>
            hasUserGesture : false
            initiator : {
                type : other
            }
            loaderId : <string>
            redirectHasExtraInfo : false
            request : {
                headers : {
                    Accept-Language : en-us,en;q=0.9
                    Upgrade-Insecure-Requests : 1
                    User-Agent : <string>
                    sec-ch-ua : "content_shell";v="999"
                    sec-ch-ua-mobile : ?0
                    sec-ch-ua-platform : "Unknown"
                }
                initialPriority : VeryHigh
                isSameSite : true
                method : GET
                mixedContentType : none
                referrerPolicy : strict-origin-when-cross-origin
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/simple-prerender.html
            }
            requestId : <string>
            timestamp : <number>
            type : Document
            wallTime : <number>
        }
        sessionId : <string>
    }
    [1] : {
        method : Network.requestWillBeSent
        params : {
            documentURL : <string>
            frameId : <string>
            hasUserGesture : false
            initiator : {
                type : script
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/simple-prerender.html
            }
            loaderId : <string>
            redirectHasExtraInfo : false
            request : {
                headers : {
                    Accept : text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7
                    Referer : http://127.0.0.1:8000/inspector-protocol/prerender/resources/simple-prerender.html
                    Sec-Purpose : prefetch;prerender
                    Sec-Speculation-Tags : null
                    Upgrade-Insecure-Requests : 1
                    sec-ch-ua : "content_shell";v="999"
                    sec-ch-ua-mobile : ?0
                    sec-ch-ua-platform : "Unknown"
                }
                initialPriority : VeryLow
                method : GET
                referrerPolicy : strict-origin-when-cross-origin
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            requestId : <string>
            timestamp : <number>
            type : Prefetch
            wallTime : <number>
        }
        sessionId : <string>
    }
]

Running test: pipelineIdShouldDifferIfMigrated
pipelineId should differ if a prefetch is triggered and then a prerender is triggered
[
    [0] : {
        method : Preload.prefetchStatusUpdated
        params : {
            initiatingFrameId : <string>
            key : {
                action : Prefetch
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            prefetchStatus : PrefetchNotFinishedInTime
            prefetchUrl : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            requestId : <string>
            status : Running
        }
        sessionId : <string>
    }
    [1] : {
        method : Preload.prerenderStatusUpdated
        params : {
            key : {
                action : Prerender
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 1
            status : Pending
        }
        sessionId : <string>
    }
    [2] : {
        method : Preload.prerenderStatusUpdated
        params : {
            key : {
                action : Prerender
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 1
            status : Running
        }
        sessionId : <string>
    }
    [3] : {
        method : Preload.prefetchStatusUpdated
        params : {
            initiatingFrameId : <string>
            key : {
                action : Prefetch
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 0
            prefetchStatus : PrefetchResponseUsed
            prefetchUrl : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            requestId : <string>
            status : Success
        }
        sessionId : <string>
    }
    [4] : {
        method : Preload.prerenderStatusUpdated
        params : {
            key : {
                action : Prerender
                loaderId : <string>
                url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
            }
            pipelineId : <string>
            pipelineIdIndex : 1
            status : Ready
        }
        sessionId : <string>
    }
]
